Further reduce plugin size
The zip data is now saved in the comment of a function instead of a base64 string, this allows for a larger character-set to encode the data.
The js, css, html, and json files are now also minimized before being zipped.
Finally, the output of rollup itself is minimized to further reduce the size of the plugin.